requirements.txt
anthropic>=0.39.0
python-dotenv>=1.0.0
pydantic>=2.0.0
email-validator>=2.0.0
